refactor(errors): move to src/lib/errors/*
- clearer structure
- facilitate polyfilling sentry
