+++ src/actions/sso/authenticationRequest.ts
@@ -6,9 +6,8 @@ import { cancelSSO } from '.'
 import { Linking } from 'react-native'
 import { JolocomLib } from 'jolocom-lib'
 import { ThunkAction } from '../../store'
-import { AppError } from '../../lib/errors'
+import { AppError, ErrorCode } from 'src/lib/errors'
 import { keyIdToDid } from 'jolocom-lib/js/utils/helper'
-import ErrorCode from '../../lib/errorCodes'
 import { generateIdentitySummary } from './utils'
 import { AuthenticationRequestSummary } from './types'
