+++ src/lib/errors/codes.ts
@@ -1,5 +1,5 @@
 // @NOTE when you add a new ErrorCode, remember to add it to src/locales/strings
-enum ErrorCode {
+export enum ErrorCode {
   Unknown = 'Unknown',
   // actions/account/index
   WalletInitFailed = 'WalletInit',
@@ -25,5 +25,3 @@ enum ErrorCode {
   // ui/generic/appInit
   AppInitFailed = 'AppInitFailed',
 }
-
-export default ErrorCode
